% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lotvolt_comp}
\alias{lotvolt_comp}
\title{Run Lotka-Voltera comparison simulation}
\usage{
lotvolt_comp(w1, b, obs_errors = 0, u_d, u_e, burnin = 100,
  sample_scale = 1L, sigma_e = 0.03, return_ts = FALSE)
}
\arguments{
\item{w1}{Weight for species 1}

\item{b}{Competition matrix}

\item{obs_errors}{Either 0 (no observation errors) or a matrix of observation
errors}

\item{u_d}{Mean 0, standard deviation 1 demographic deviations.}

\item{u_e}{Mean 0, standard deviation 1 environmental deviations.}

\item{burnin}{How many iterations to discard as burnin.}

\item{sample_scale}{How often to sample the 'truth'. If set to 1 then all
observations are kept. If set to 2, for example, then every second
observation is discarded.}

\item{sigma_e}{Environmental standard deviation values to pass on.}

\item{return_ts}{Logical: should the time series of species biomasses be
returned or the mean-variance statistics?}
}
\description{
A wrapper function to facilitate running simulations for two species.
}

