% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lotvolt_sp}
\alias{lotvolt_sp}
\title{Run Lotka-Voltera comparison simulation for many species}
\usage{
lotvolt_sp(n_sim, b, obs_errors = 0, u_d, u_e, burnin = 100,
  ef_interval = 20)
}
\arguments{
\item{n_sim}{Number of Monte Carlo simulations to run}

\item{b}{Competition value (same for all species)}

\item{obs_errors}{Either 0 (no observation errors) or a matrix of observation
errors}

\item{u_d}{Mean 0, standard deviation 1 demographic deviations.}

\item{u_e}{Mean 0, standard deviation 1 environmental deviations.}

\item{burnin}{How many iterations to discard as burn in.}

\item{ef_interval}{How many bins to divide the efficient frontier into when
calculating the frontier based on the Monte Carlo results.}
}
\description{
A wrapper function to facilitate running simulations where the number of
species can be more than 2.
}

