% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_frontiers}
\alias{plot_frontiers}
\title{Plot efficient frontiers}
\usage{
plot_frontiers(dat, xlim = NULL, ylim = NULL, show = seq_along(dat),
  colour_lines = FALSE, col = NULL, add = FALSE, lwd = 3, ...)
}
\arguments{
\item{dat}{A list of data frames containing columsn for variability (\code{v})
and mean (\code{m}).}

\item{xlim}{Possible x limits}

\item{ylim}{Possible y limits}

\item{show}{Vector of portfolios to show}

\item{colour_lines}{Should the lines be coloured?}

\item{col}{A vector of colours for the lines.}

\item{add}{Should lines be added to an existing plot?}

\item{lwd}{Line width}

\item{...}{Any extra arguments to pass to \code{plot.default}.}
}
\description{
Plot efficient frontiers
}

